+++ debug_toolbar/toolbar.py
@@ -93,14 +93,7 @@ class DebugToolbar:
 
         If False, the panels will be loaded via Ajax.
         """
-        if (render_panels := self.config["RENDER_PANELS"]) is None:
-            # If wsgi.multiprocess isn't in the headers, then it's likely
-            # being served by ASGI. This type of set up is most likely
-            # incompatible with the toolbar until
-            # https://github.com/jazzband/django-debug-toolbar/issues/1430
-            # is resolved.
-            render_panels = self.request.META.get("wsgi.multiprocess", True)
-        return render_panels
+        return self.config["RENDER_PANELS"] or False
 
     # Handle storing toolbars in memory and fetching them later on